Updated debank types
