--- lambdas/src/shared/debank.ts
@@ -1,55 +1,0 @@
-import { NetworkNames } from './domain-types'
-
-export type ProtocolAsset = {
-  id: string
-  chain: string
-  name: string
-  site_url: string
-  logo_url: string
-  has_supported_portfolio: boolean
-  tvl: number
-  net_usd_value: number
-  asset_usd_value: number
-  debt_usd_value: number
-}
-
-export type WalletAsset = {
-  usd_value: number
-}
-
-export type DebankToken = {
-  id: string
-  chain: string
-  name: string
-  symbol: string
-  display_symbol: string
-  optimized_symbol: string
-  decimals: number
-  logo_url: string
-  protocol_id: string
-  price: number
-  price_24h_change: number
-  is_verified: boolean
-  is_core: boolean
-  is_wallet: boolean
-  time_at: number
-  amount: number
-  raw_amount: number
-  raw_amount_hex_str: string
-}
-
-export enum DebankNetworkNames {
-  ethereumMainnet = 'eth',
-  arbitrumMainnet = 'arb',
-  polygonMainnet = 'matic',
-  optimismMainnet = 'op',
-  baseMainnet = 'base',
-}
-
-export const DebankNetworkNameToOurs = {
-  [DebankNetworkNames.ethereumMainnet]: NetworkNames.ethereumMainnet,
-  [DebankNetworkNames.arbitrumMainnet]: NetworkNames.arbitrumMainnet,
-  [DebankNetworkNames.polygonMainnet]: NetworkNames.polygonMainnet,
-  [DebankNetworkNames.optimismMainnet]: NetworkNames.optimismMainnet,
-  [DebankNetworkNames.baseMainnet]: NetworkNames.baseMainnet,
-}
